Make RemoteKernelManager more independent
Make RemoteKernelManager (RKM) able to be run without a (grand)parent
EnterpriseGatewayApp (EGA) instance so it can be used by nbclient.

The config which RKM with EGA was being referred to using traitlet
lineage like `self.parent.parent.property_name`. Change both to inherit
from a Configurable mixin which contains all of EGA's previous config.
Link the attributes of the RKM instance with EGA if available to keep
old behaviour.

Modify other properties RKM uses that are not traits to become
`@property`s which fall back to sane defaults if running independently.

Change RKM to be able to generate a kernel id if necessary. Change
ProcessProxy to use provided kernel ids. Move kernel id generation
logic out of RemoteMappingKernelManager.

Resolves #803
